Added partial download + Fixed reviews and text dialog + Fixed some TTS issues

+++ app/src/main/java/com/lagradost/quicknovel/ui/download/AnyAdapter.kt
@@ -228,7 +228,7 @@ class AnyAdapter(
                         "${card.downloadedCount}/${card.downloadedTotal}" + if (card.ETA == "") "" else " - ${card.ETA}"
 
                     downloadProgressbar.apply {
-                        max = card.downloadedTotal * 100
+                        max = card.downloadedTotal.toInt() * 100
 
                         // shitty check for non changed
                         if (same) {
@@ -236,7 +236,7 @@ class AnyAdapter(
                                 this,
                                 "progress",
                                 progress,
-                                card.downloadedCount * 100
+                                card.downloadedCount.toInt() * 100
                             )
 
                             animation.duration = 500
@@ -244,7 +244,7 @@ class AnyAdapter(
                             animation.interpolator = DecelerateInterpolator()
                             animation.start()
                         } else {
-                            progress = card.downloadedCount * 100
+                            progress = card.downloadedCount.toInt() * 100
                         }
                         //download_progressbar.progress = card.downloadedCount
                         alpha = if (card.downloadedCount >= card.downloadedTotal) 0f else 1f
